v5
added a slew of features

I'm thinking of not exporting a xlsx until after the tail end of the visual graph. It's a very expensive [computationally] operation to export to xlsx and I've found a new way to "pickle" the data (idk why the name, but it means to serialize the data into a binary format) and I can use that method to pass it to the "analysis" portion of code and "unpack" it there fairly quickly meaning you can focus on the inferences while it's still exporting.

I want to expand the analysis from 2 years to 3-5 years. This won't affect any of the calculations, it will make an export of the data to excel much larger though. A filter will still be in place to only select those that have at least 2 years of data, but this is for preparation for my next step.

I want to add "seasonal index" calculations for the sectors. I default to quarterly because I like the nice round quarter numbers (i.e. 63 trading days) and most businesses seem to plan around that. I could do a monthly, but it would be a bastardized month (21 trading days != the same length). Or I could try weekly index's (but that would be noisy), as 52 weeks in a year is nice and smooth, but quarterly seems to be the nice midway point between those choices.

This way when you look at a stock and consider buying it, you can see how the historical performance of the sector has been and make a quick call if there will be a seasonal downturn (I notice energy sector dips in the summer and rise in the winter)

I've also added a 252 day seasonal forecast (excessive, but if I reduce the data to quarterly and then forecast out, it chops off any data after the last full quarter, by choosing 252 it focuses on the past 2 years of data, which is why I want to extend the analysis to 5 years because it will be more accurate). (2 years might be just fine)

I also want to calculate quality metrics and give an estimate about how accurate the forecast is.

As I've been doing since yesterday,, I will upload the report to google drive as a pdf so you can see visually all this stuff I'm talking about (without the code)