added stop loss price
week cuts off on Sunday.  So if running this on friday, it will cut the date to last sunday. Not sure how to fix this, but it means forecasts will not take last trading week into consideration (i.e. it's not trained on the last data point).  Which means there is a type of holdout on the last week that can be used if the stock is moving in the right direction within the prediction interval (i.e. stop loss price is 50% inbetween last asfreq price and inital prediction interval).