v10.2 updated forecast plots
I saw a discrepency between the way fbprophet forecasts and ets forecasts were plotting their outcome variable. So I decided to replot fbprophet using some different parameter's to tease out the issue which

I'm still not certain how to aptly describe the discrepency, but I can plot it.

Problem:
For fbprophet, the mean yhat is occasionally out of bounds of both confidence intervals (I've been noticing it on the latest value for recent forecasts), and this problem doesn't happen w ETS forecasts.

Top plot forecast is ETS but doesn't list prior forecasts (making it hard to do a 1:1 comparison).

I'm thinking I need to do some more investigating, but my working theories on the issue are

ETS and fbprophet are using an unknown combination of confidence vs prediction intervals (I presume ETS is prediction based on terms used in the documentation) and their not matching with actual results.

Or fbprophet is plotting as a mean confidence interval (not prediction) and the latest price is simply out of the bound of the means of the average, not a specific prediction forecast.